Fixed #17788 -- Added a batch_size argument for qs.bulk_create()
The qs.bulk_create() method did not work with large batches together
with SQLite3. This commit added a way to split the bulk into smaller
batches. The default batch size is unlimited except for SQLite3 where
the batch size is limited to 1000 SQL parameters per batch.

Thanks to everybody who participated in the discussion at Trac.
Special thanks to alpar for the report.