Added a failing test for deferred model .save()
When saving a deferred model, the meta class remember which fields are
loaded, which not. This causes problems when setting some field to a
value, as now the meta class remembers that the field is non-deferred.
Of course, the correct approach would be to have the remembering done
per instance instead.

Added a test showing the above mentioned problem.