Fixed #18375 -- Ensured F-expressions use existing joins consistently
Previously a .filter(revfk__somecol=F('revfk__anothercol')) created two
joins, one for the F() expression, another for the revfk__somecol
lookup. Fixed this in a way that the F() expression will target the
lookup's join in this case. This is particularly important if there
were multiple lookups in one .filter() call, as then what the F()
targets could be kwargs order dependant, and thus dict ordering
dependant.